[#749] Move fancy stuff from headernav to component
- Consolidates upgrade modal and logic into FancyIcon for reusability
- Simplifies HeaderUserNav by delegating modal handling to FancyIcon
- Enhances maintainability and reduces code duplication
